Use Mixins for simplification
- Split Sw360Error in independent class
- Split the main codebase in Mixins to reduce size of main source code
This makes clearer and easier to update specific sw360 areas
