+++ pyproject.toml
@@ -31,6 +31,7 @@ flake8 = ">=3.7.8"
 responses = "^0.12.1"
 pytest = "^6.1.2"
 coverage = "^5.3"
+types-requests = "^2.27.11"
 
 [build-system]
 requires = ["poetry>=0.12"]
